refactor(test): app 모듈 중복 테스트 지원 클래스를 test-support 모듈로 통합
- 8개 순수 중복 클래스 삭제: TestContainerSupport, TestStompSession,
  TypeReferenceFrameHandler, MessageResponse, ServiceTest,
  ServiceTestConfig, ShutDownTestScheduler, TestTaskScheduler
- 40개 파일의 import를 coffeeshout.support.* 경로로 교체
- app 전용 유지: IntegrationTestConfig(실 FlowScheduler),
  support.test.IntegrationTest(@Transactional 없음),
  fixture.IntegrationTestSupport(E2E용)
