--- backend/app/src/test/java/coffeeshout/global/MessageResponse.java
@@ -1,4 +1,0 @@
-package coffeeshout.global;
-
-public record MessageResponse(long duration, String payload) {
-}
