--- backend/app/src/test/java/coffeeshout/global/ServiceTest.java
@@ -1,20 +1,0 @@
-package coffeeshout.global;
-
-import coffeeshout.fixture.TestContainerSupport;
-import coffeeshout.global.config.ServiceTestConfig;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.context.annotation.Import;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.bean.override.mockito.MockitoBean;
-import org.springframework.transaction.annotation.Transactional;
-
-@SpringBootTest
-@Import(ServiceTestConfig.class)
-@ActiveProfiles("test")
-@Transactional
-public abstract class ServiceTest extends TestContainerSupport {
-
-    @MockitoBean
-    protected ApplicationEventPublisher eventPublisher;
-}
