+++ backend/app/src/test/java/coffeeshout/websocket/WebSocketConnectionTest.java
@@ -3,7 +3,7 @@ package coffeeshout.websocket;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import coffeeshout.fixture.TestStompSession;
+import coffeeshout.support.TestStompSession;
 import coffeeshout.fixture.WebSocketIntegrationTestSupport;
 import java.util.concurrent.ExecutionException;
 import org.junit.jupiter.api.Test;
